feat(build): add tui shortcut recipe to justfile
Add `just tui` as a convenient alias for `just view-tests`.
